Article_generator_chatbot/requirements.txt
streamlit==1.47.0
python-dotenv==1.1.1
google-generativeai==0.8.5
requests==2.32.4
